backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
azure-storage-blob==12.22.0
azure-identity==1.18.0
openai==1.51.0
python-pptx==1.0.2
openpyxl==3.1.5
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.27.2
requests==2.32.3
